Extract attendance section into POSBookingAttendanceSectionView
Move attendance UI, state management, button style, and localization
out of POSBookingDetailView into a dedicated file.
