+++ Modules/Sources/PointOfSale/Presentation/Bookings/POSBookingDetailView.swift
@@ -1,6 +1,5 @@
 import SwiftUI
 import struct Yosemite.POSBooking
-import enum Yosemite.BookingAttendanceStatus
 
 struct POSBookingDetailView: View {
     let booking: POSBooking
@@ -16,9 +15,6 @@ struct POSBookingDetailView: View {
     @State private var cancelModalState: CancelBookingModalState?
     @State private var showPaymentView = false
     @State private var paymentModel: POSPaymentModel?
-    @State private var isUpdatingAttendance = false
-    @State private var attendanceUpdateError = false
-
 
     private var actionTintColor: Color {
         colorScheme == .dark ? .posSecondary : .posPrimaryContainer
@@ -86,7 +82,7 @@ struct POSBookingDetailView: View {
             ScrollView {
                 VStack(alignment: .leading, spacing: POSSpacing.large) {
                     bookingDetailsSection
-                    attendanceSection
+                    POSBookingAttendanceSectionView(booking: booking)
                     customerSection
                     paymentBreakdownSection
                     bookingNoteSection
@@ -271,61 +267,6 @@ struct POSBookingDetailView: View {
         }
     }
 
-    // MARK: - Attendance Status
-
-    @ViewBuilder
-    private var attendanceSection: some View {
-        VStack(alignment: .leading, spacing: POSSpacing.small) {
-            sectionTitleWithAction(title: Localization.attendanceStatusTitle) {
-                if isUpdatingAttendance {
-                    ProgressView()
-                        .frame(maxHeight: .infinity, alignment: .center)
-                } else {
-                    HStack(spacing: POSSpacing.small) {
-                        attendanceButton(Localization.attendedPill,
-                                         isSelected: booking.attendanceDisplay == .attended) {
-                            performAttendanceUpdate(targetStatus: .attended)
-                        }
-                        attendanceButton(Localization.unattendedPill,
-                                         isSelected: booking.attendanceDisplay == .unattended) {
-                            performAttendanceUpdate(targetStatus: .unattended)
-                        }
-                    }
-                }
-            }
-            .sectionCard()
-
-            if attendanceUpdateError {
-                Text(Localization.attendanceUpdateErrorMessage)
-                    .font(.posBodySmallRegular())
-                    .foregroundStyle(Color.posError)
-                    .padding(.horizontal, POSPadding.xSmall)
-            } else {
-                Text(Localization.attendanceSubtitle)
-                    .font(.posBodySmallRegular())
-                    .foregroundStyle(Color.posOnSurfaceVariantHighest)
-                    .padding(.horizontal, POSPadding.xSmall)
-            }
-        }
-    }
-
-    private func performAttendanceUpdate(targetStatus: BookingAttendanceStatus) {
-        guard targetStatus != booking.attendanceStatus else { return }
-        attendanceUpdateError = false
-        isUpdatingAttendance = true
-        Task { @MainActor in
-            do {
-                try await bookingsModel.bookingsController.updateAttendanceStatus(
-                    bookingID: booking.id,
-                    status: targetStatus
-                )
-            } catch {
-                attendanceUpdateError = true
-            }
-            isUpdatingAttendance = false
-        }
-    }
-
     // MARK: - Payment Breakdown
 
     private var paymentBreakdownSection: some View {
@@ -398,19 +339,6 @@ struct POSBookingDetailView: View {
     }
 
     @ViewBuilder
-    private func attendanceButton(_ title: String, isSelected: Bool, action: @escaping () -> Void) -> some View {
-        if isSelected {
-            Button(title, action: action)
-                .buttonStyle(POSFilledAttendanceButtonStyle())
-                .accessibilityAddTraits(.isSelected)
-        } else {
-            Button(title, action: action)
-                .buttonStyle(POSOutlinedButtonStyle(size: .compact))
-                .accessibilityRemoveTraits(.isSelected)
-        }
-    }
-
-    @ViewBuilder
     private func detailRow(label: String, value: String) -> some View {
         HStack {
             Text(label)
@@ -442,28 +370,6 @@ struct POSBookingSectionCardModifier: ViewModifier {
 extension View {
     func sectionCard() -> some View {
         modifier(POSBookingSectionCardModifier())
-    }
-}
-
-// MARK: - Attendance Button Style
-
-private struct POSFilledAttendanceButtonStyle: ButtonStyle {
-    @Environment(\.colorScheme) private var colorScheme
-
-    private var fillColor: Color {
-        colorScheme == .dark ? .posSecondary : .posPrimaryContainer
-    }
-
-    func makeBody(configuration: Configuration) -> some View {
-        configuration.label
-            .font(.posBodySmallBold())
-            .padding(.vertical, POSPadding.small)
-            .padding(.horizontal, POSPadding.medium)
-            .foregroundStyle(Color.posOnInverseSurface)
-            .background(fillColor)
-            .clipShape(RoundedRectangle(cornerRadius: POSCornerRadiusStyle.small.value))
-            .opacity(configuration.isPressed ? 0.7 : 1.0)
-            .animation(.easeOut(duration: 0.15), value: configuration.isPressed)
     }
 }
 
@@ -537,30 +443,6 @@ private enum Localization {
         comment: "Subtitle text below the booking note section explaining the note is private."
     )
 
-    static let attendanceStatusTitle = NSLocalizedString(
-        "pos.bookingDetailView.attendanceStatusTitle",
-        value: "Attendance status",
-        comment: "Section title for the attendance status in booking details."
-    )
-
-    static let attendedPill = NSLocalizedString(
-        "pos.bookingDetailView.attendedPill",
-        value: "Attended",
-        comment: "Label for the attended pill button in booking attendance section."
-    )
-
-    static let unattendedPill = NSLocalizedString(
-        "pos.bookingDetailView.unattendedPill",
-        value: "Unattended",
-        comment: "Label for the unattended pill button in booking attendance section."
-    )
-
-    static let attendanceSubtitle = NSLocalizedString(
-        "pos.bookingDetailView.attendanceSubtitle",
-        value: "Mark attendance to keep your reports accurate and spot booking trends.",
-        comment: "Subtitle text below the attendance section explaining why marking attendance matters."
-    )
-
     static let paymentTitle = NSLocalizedString(
         "pos.bookingDetailView.paymentTitle",
         value: "Payment",
@@ -629,11 +511,6 @@ private enum Localization {
         comment: "Menu action to cancel a booking from the POS booking detail view."
     )
 
-    static let attendanceUpdateErrorMessage = NSLocalizedString(
-        "pos.bookingDetailView.attendanceUpdateError",
-        value: "Failed to update attendance. Please try again.",
-        comment: "Error message shown below attendance section when updating attendance status fails."
-    )
 }
 
 // MARK: - Previews
